refactor(resources): move components to _components
